+++ lib/redcarpet_camo_renderer.rb
